error checking added to update

+++ enable-enhanced-session-mode/ubuntu/install24_04.sh
@@ -13,6 +13,10 @@ fi
 
 # Make sure system is in good shape
 apt update
+if [ $? -ne 0 ]; then
+    echo "Error on updating"
+    exit 1
+fi
 apt upgrade -y
 apt autoremove
